fix(web): lazy-load Leaflet to fix SSR window-is-not-defined crash
